# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/file/getfilelist": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取用户文件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "currentPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "filePath",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "fileType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageCount",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_models.RespData"
                        }
                    },
                    "400": {
                        "description": "参数出错",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/filetransfer/batchDownloadFile": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "文件批量下载接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "多个用户文件标识符，以逗号隔开",
                        "name": "userFileIds",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cookie",
                        "name": "cookie",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器响应成功，根据响应code判断是否成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "参数出错",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/filetransfer/downloadfile": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "文件单个下载接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "单个用户文件标识符",
                        "name": "userFileId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cookie",
                        "name": "cookie",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器响应成功，根据响应code判断是否成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "参数出错",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/filetransfer/getstorage": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取用户存储容量",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api_models.RespData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api_models.UserStorageReqAPI"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "cookie校验失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/recoveryfile/deleterecoveryfile": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除多个回收站文件",
                "parameters": [
                    {
                        "description": "用户文件id",
                        "name": "userFileId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Cookie",
                        "name": "cookie",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_models.RespData"
                        }
                    },
                    "400": {
                        "description": "参数出错",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/recoveryfile/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取回收站文件列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cookie",
                        "name": "cookie",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "服务器响应成功，根据响应code判断是否成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api_models.RespDataList"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "datalist": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api_models.RecoveryListRespAPI"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数出错",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_models.RecoveryListRespAPI": {
            "type": "object",
            "properties": {
                "deleteBatchNum": {
                    "type": "string"
                },
                "deleteTime": {
                    "type": "string"
                },
                "extendName": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "fileType": {
                    "type": "integer"
                },
                "isDir": {
                    "type": "integer"
                },
                "uploadTime": {
                    "type": "string"
                },
                "userFileId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "api_models.RespData": {
            "type": "object",
            "required": [
                "code",
                "data",
                "message",
                "success"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api_models.RespDataList": {
            "type": "object",
            "required": [
                "code",
                "dataList",
                "message",
                "success",
                "total"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "dataList": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api_models.UserStorageReqAPI": {
            "type": "object",
            "properties": {
                "storageSize": {
                    "type": "integer"
                },
                "totalStorageSize": {
                    "type": "integer"
                }
            }
        }
    }
}